refactor(DropdownMenu): move CheckIcon in own file
