+++ packages/wedges/src/components/DropdownMenu/DropdownMenu.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 
 import { cn } from "../../helpers/utils";
+import { CheckIcon } from "../icons/CheckIcon";
 import { Kbd, type KbdElement, type KbdProps } from "../Kbd";
 
 /* ---------------------------------- Types --------------------------------- */
@@ -126,25 +127,6 @@ const DropdownMenuSeparator = React.forwardRef<
 /* -------------------------------------------------------------------------- */
 /*                                Checkbox Item                               */
 /* -------------------------------------------------------------------------- */
-const CheckIcon = ({ className, ...otherProps }: React.ComponentProps<"svg">) => (
-  <svg
-    {...otherProps}
-    className={cn("!opacity-100", className)}
-    fill="none"
-    height="24"
-    stroke="currentColor"
-    viewBox="0 0 24 24"
-    width="24"
-  >
-    <path
-      d="M7.75 12.75L10 15.25L16.25 8.75"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="1.5"
-    />
-  </svg>
-);
 
 const DropdownMenuCheckboxItem = React.forwardRef<
   React.ElementRef<typeof DropdownMenuPrimitive.CheckboxItem>,
@@ -244,7 +226,7 @@ const DropdownMenuSubContent = React.forwardRef<
         "data-[side=bottom]:animate-wg-fade-in-down data-[side=left]:animate-wg-fade-in-left data-[side=right]:animate-wg-fade-in-right data-[side=top]:animate-wg-fade-in-up data-[state=closed]:animate-wg-fade-out",
 
         // base styles
-        "z-50 flex min-w-[9rem] origin-[var(--radix-popper-transform-origin)] flex-col gap-2 rounded-lg bg-white py-2 shadow-wg-overlay dark:border dark:border-surface dark:bg-neutral-800 dark:shadow-none",
+        "z-50 flex min-w-36 origin-[var(--radix-popper-transform-origin)] flex-col gap-2 rounded-lg bg-white py-2 shadow-wg-overlay dark:border dark:border-surface dark:bg-neutral-800 dark:shadow-none",
 
         // has checkbox or radio item - offset start padding
         inset && "[--wg-offset-padding-left:34px]",
